requirements.txt
google-cloud-translate>=2.0.1
Django>=3.1.4
psycopg2-binary
nltk>=3.5